Changed shader compilation to use uint8_t instead of uint32_t for the binary code.
- Added build metal-cpp file.
- Minor clean up.

+++ Humble2/src/Humble2/Resources/TypeDescriptors.h
@@ -150,7 +150,7 @@ namespace HBL2
 		ShaderType type = ShaderType::RASTERIZATION;
 		struct ShaderStage
 		{
-			Span<const uint32_t> code;
+			Span<const uint8_t> code;
 			const char* entryPoint;
 		};
 		ShaderStage VS;
